+++ src/loader/mod.rs
@@ -1,0 +1,26 @@
+/*
+ * This file is part of Twig (ported to Rust).
+ *
+ * For the copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+/**
+ * Twig library for rust
+ *
+ * @author Colin Kiegel <kiegel@gmx.de>
+ */
+
+/////////////
+// imports //
+/////////////
+
+
+
+/////////////
+// exports //
+////////////
+
+pub mod api;
+pub mod array;
+pub mod filesystem;